requirements.txt
alembic==1.15.2
asyncpg==0.30.0
bcrypt==4.0.1
fastapi[all]
passlib[bcrypt]==1.7.4
pydantic-settings==2.8.1
python-jose==3.4.0
requests==2.32.3
SQLAlchemy==2.0.40
